+++ .gitignore
@@ -1,3 +1,6 @@
 .venv
 venv
 .qodo
+
+**/__pycache__/
+**/*.pyc
